implement delete functionaly for resource

+++ src/main/java/com/sagarandcompany/linkSharing/repository/ResourceRatingRepository/ResourceRatingRepoImpl.java
@@ -1,14 +1,10 @@
 package com.sagarandcompany.linkSharing.repository.ResourceRatingRepository;
 
-import com.sagarandcompany.linkSharing.domains.LinkResource;
-import com.sagarandcompany.linkSharing.domains.Resource;
-import com.sagarandcompany.linkSharing.domains.ResourceRating;
-import com.sagarandcompany.linkSharing.domains.User;
+import com.sagarandcompany.linkSharing.domains.*;
 import com.sagarandcompany.linkSharing.utility.ResourceRatingDTO;
 import com.sagarandcompany.linkSharing.utility.ResponseDTO;
-import org.hibernate.Query;
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
+import org.hibernate.*;
+import org.hibernate.criterion.Restrictions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
@@ -27,8 +23,8 @@ public class ResourceRatingRepoImpl implements ResorceRatingRepo {
 
         User user = getSession().get(User.class, User.getLoginUser().getUser_id());
         Resource resource = getSession().get(Resource.class, resourceRatingDTO.getResource_id());
-        resourceRating.setUser(user);
-        resourceRating.setResource(resource);
+        resourceRating.setUserId(user.getUser_id());
+        resourceRating.setResourceId(resource.getResource_id());
 
         getSession().save(resourceRating);
         if (resourceRating.getResource_rating_id() != null)
@@ -38,27 +34,42 @@ public class ResourceRatingRepoImpl implements ResorceRatingRepo {
 
     }
 
-    public ResourceRating get(Long id)
-    {
-        if(id!=null) {
+    public ResourceRating get(Long id) {
+        if (id != null) {
             ResourceRating resourceRating = getSession().get(ResourceRating.class, id);
             if (resourceRating != null) {
                 return resourceRating;
             } else
                 return null;
-        }
-        else
+        } else
             return null;
     }
 
-     public Boolean delete(Long id)
-     {
-        Session session=getSession();
-        ResourceRating resourceRating=getSession().get(ResourceRating.class,id);
-            session.delete(resourceRating);
-            return true;
+    public Boolean delete(Long id) {
+        Session session = getSession();
+        ResourceRating resourceRating = getSession().get(ResourceRating.class, id);
+        session.delete(resourceRating);
+        return true;
 
-     }
+    }
+
+    public Boolean deleteByResource(Resource resource) {
+        Session session = getSession();
+        Criteria criteria = session.createCriteria(ResourceRating.class);
+        criteria.add(Restrictions.eq("resourceId", resource.getResource_id()));
+        criteria.add(Restrictions.eq("userId", User.getLoginUser().getUser_id()));
+        criteria.setMaxResults(1);
+        ResourceRating resourceRating = (ResourceRating) criteria.uniqueResult();
+
+        if (resourceRating != null) {
+            session.delete(resourceRating);
+            session.flush();
+            return true;
+        }
+        return false;
+
+    }
+
     private Session getSession() {
         Session session = sessionFactory.getCurrentSession();
         return session;
